#157 Feat: 북마크-태크 컨벤션 및 카테고리 컨벤션 일부 완료
