--- linknamu/src/main/java/com/kakao/linknamu/bookmarkTag/dto/CreateBookmarkTagRequestDto.java
@@ -1,6 +1,0 @@
-package com.kakao.linknamu.bookmarkTag.dto;
-
-public record CreateBookmarkTagRequestDto(
-        String tagName
-) {
-}
