--- linknamu/src/main/java/com/kakao/linknamu/bookmarkTag/dto/DeleteBookmarkTagRequestDto.java
@@ -1,7 +1,0 @@
-package com.kakao.linknamu.bookmarkTag.dto;
-
-public record DeleteBookmarkTagRequestDto(
-        Long bookmarkId,
-        Long tagId
-) {
-}
