+++ linknamu/src/main/java/com/kakao/linknamu/bookmarktag/BookmarkTagExceptionStatus.java
@@ -1,7 +1,6 @@
-package com.kakao.linknamu.bookmarkTag;
+package com.kakao.linknamu.bookmarktag;
 
 import com.kakao.linknamu.core.exception.BaseExceptionStatus;
-
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
