+++ linknamu/src/main/java/com/kakao/linknamu/bookmarktag/dto/CreateBookmarkTagRequestDto.java
@@ -1,0 +1,6 @@
+package com.kakao.linknamu.bookmarktag.dto;
+
+public record CreateBookmarkTagRequestDto(
+	String tagName
+) {
+}
