+++ linknamu/src/main/java/com/kakao/linknamu/bookmarktag/dto/DeleteBookmarkTagRequestDto.java
@@ -1,0 +1,7 @@
+package com.kakao.linknamu.bookmarktag.dto;
+
+public record DeleteBookmarkTagRequestDto(
+	Long bookmarkId,
+	Long tagId
+) {
+}
