+++ linknamu/src/main/java/com/kakao/linknamu/bookmarktag/repository/BookmarkTagJpaRepositoryCustom.java
@@ -1,10 +1,11 @@
-package com.kakao.linknamu.bookmarkTag.repository;
+package com.kakao.linknamu.bookmarktag.repository;
 
 import com.kakao.linknamu.bookmark.dto.BookmarkSearchCondition;
 import com.kakao.linknamu.bookmark.entity.Bookmark;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-public interface BookmarkTagJPARepositoryCustom {
-    Page<Bookmark> search(BookmarkSearchCondition condition, Long userId, Pageable pageable);
+public interface BookmarkTagJpaRepositoryCustom {
+	Page<Bookmark> search(BookmarkSearchCondition condition, Long userId, Pageable pageable);
 }
+
