+++ linknamu/src/main/java/com/kakao/linknamu/bookmarktag/service/BookmarkTagReadService.java
@@ -1,24 +1,22 @@
-package com.kakao.linknamu.bookmarkTag.service;
+package com.kakao.linknamu.bookmarktag.service;
 
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-
-import com.kakao.linknamu.bookmarkTag.BookmarkTagExceptionStatus;
-import com.kakao.linknamu.bookmarkTag.entity.BookmarkTag;
-import com.kakao.linknamu.bookmarkTag.entity.BookmarkTagId;
-import com.kakao.linknamu.bookmarkTag.repository.BookmarkTagJPARepository;
+import com.kakao.linknamu.bookmarktag.BookmarkTagExceptionStatus;
+import com.kakao.linknamu.bookmarktag.entity.BookmarkTag;
+import com.kakao.linknamu.bookmarktag.entity.BookmarkTagId;
+import com.kakao.linknamu.bookmarktag.repository.BookmarkTagJpaRepository;
 import com.kakao.linknamu.core.exception.Exception400;
 import com.kakao.linknamu.core.exception.Exception404;
 import com.kakao.linknamu.tag.entity.Tag;
-
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @RequiredArgsConstructor
 @Service
 public class BookmarkTagReadService {
 
-	private final BookmarkTagJPARepository bookmarkTagJPARepository;
+	private final BookmarkTagJpaRepository bookmarkTagJPARepository;
 
 	public List<Tag> findTagByBookmarkId(Long bookmarkId) {
 		return bookmarkTagJPARepository.findTagByBookmarkId(bookmarkId);
