+++ linknamu/src/main/java/com/kakao/linknamu/bookmarktag/service/BookmarkTagSearchService.java
@@ -1,0 +1,36 @@
+package com.kakao.linknamu.bookmarktag.service;
+
+import com.kakao.linknamu.bookmark.dto.BookmarkSearchCondition;
+import com.kakao.linknamu.bookmark.entity.Bookmark;
+import com.kakao.linknamu.bookmarktag.repository.BookmarkTagJpaRepository;
+import com.kakao.linknamu.tag.entity.Tag;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+@Service
+public class BookmarkTagSearchService {
+	private final BookmarkTagJpaRepository bookmarkTagJPARepository;
+
+	public List<String> searchTagNamesByBookmarkId(Long bookmarkId) {
+		return bookmarkTagJPARepository.findTagNamesByBookmarkId(bookmarkId);
+	}
+
+	public List<Long> searchTagIdsByBookmarkId(Long bookmarkId) {
+		return bookmarkTagJPARepository.findTagIdsByBookmarkId(bookmarkId);
+	}
+
+	public List<Tag> findTagsByBookmarkId(Long bookmarkId) {
+		return bookmarkTagJPARepository.findTagByBookmarkId(bookmarkId);
+	}
+
+	public Page<Bookmark> search(BookmarkSearchCondition condition, Long userId, Pageable pageable) {
+		return bookmarkTagJPARepository.search(condition, userId, pageable);
+	}
+}
