+++ linknamu/src/main/java/com/kakao/linknamu/category/dto/CategorySaveRequestDto.java
@@ -3,7 +3,7 @@ package com.kakao.linknamu.category.dto;
 import jakarta.validation.constraints.NotNull;
 
 public record CategorySaveRequestDto(
-        @NotNull String categoryName,
-        @NotNull Long workspaceId
+	@NotNull String categoryName,
+	@NotNull Long workspaceId
 ) {
 }
