+++ linknamu/src/main/java/com/kakao/linknamu/category/dto/CategoryUpdateRequestDto.java
@@ -3,5 +3,6 @@ package com.kakao.linknamu.category.dto;
 import jakarta.validation.constraints.NotNull;
 
 public record CategoryUpdateRequestDto(
-        @NotNull String categoryName
-) {}
+	@NotNull String categoryName
+) {
+}
