+++ linknamu/src/main/java/com/kakao/linknamu/category/entity/Category.java
@@ -1,41 +1,28 @@
 package com.kakao.linknamu.category.entity;
 
-import java.util.Objects;
-
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-
 import com.kakao.linknamu.core.util.AuditingEntity;
 import com.kakao.linknamu.workspace.entity.Workspace;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
+import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.OnDelete;
+import org.hibernate.annotations.OnDeleteAction;
+
+import java.util.Objects;
 
 @Getter
 @NoArgsConstructor
 @Entity
 @Table(
 	name = "category_tb",
-	uniqueConstraints = {
-		@UniqueConstraint(
-			name = "workspace_categoryName unique constraint",
-			columnNames = {
-				"workspace_id",
-				"category_name"
-			}
-		)
-	}
+	uniqueConstraints =
+		{
+			@UniqueConstraint(
+				name = "workspace_categoryName unique constraint",
+				columnNames = {"workspace_id", "category_name"}
+			)
+		}
 )
 public class Category extends AuditingEntity {
 
@@ -64,12 +51,14 @@ public class Category extends AuditingEntity {
 	}
 
 	@Override
-	public boolean equals(Object o) {
-		if (this == o)
+	public boolean equals(Object obj) {
+		if (this == obj) {
 			return true;
-		if (o == null || getClass() != o.getClass())
+		}
+		if (obj == null || getClass() != obj.getClass()) {
 			return false;
-		Category category1 = (Category)o;
+		}
+		Category category1 = (Category) obj;
 		return Objects.equals(categoryId, category1.categoryId);
 	}
